move json loaders to loaders module
