+++ src/jsonmaestro/loaders.pyi
@@ -1,0 +1,9 @@
+from typing import Any, Dict
+
+
+def load_jsonc(file_path: str) -> Dict[str, Any]:
+	...
+
+
+def load_json(file_path: str) -> Dict[str, Any]:
+	...
